Rename tasks file
