+++ hardhat.config.js
@@ -13,7 +13,7 @@ if (!process.env.TEST === "true") {
   require("hardhat-gas-reporter");
 }
 require("./tasks/deploy");
-require("./tasks/deployTributeERC721");
+require("./tasks/tributeERC721Tasks");
 require("./signers");
 
 module.exports = {
